+++ include/unity/scopes/internal/ScopeImpl.h
@@ -81,6 +81,9 @@ public:
                                    ActionMetadata const& hints,
                                    PreviewListenerBase::SPtr const& reply) override;
 
+    virtual ChildScopeList child_scopes_ordered() override;
+    virtual bool set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered) override;
+
     static ScopeProxy create(MWScopeProxy const& mw_proxy, std::string const& scope_id);
 
 private: