+++ include/unity/scopes/internal/ScopeObject.h
@@ -76,6 +76,9 @@ public:
                                      MWReplyProxy const& reply,
                                      InvokeInfo const& info) override;
 
+    virtual ChildScopeList child_scopes_ordered() const override;
+    virtual bool set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered) override;
+
     virtual bool debug_mode() const override;
 
 private: