+++ include/unity/scopes/internal/ScopeObjectBase.h
@@ -21,6 +21,7 @@
 #include <unity/scopes/internal/AbstractObject.h>
 #include <unity/scopes/internal/MWQueryCtrlProxyFwd.h>
 #include <unity/scopes/internal/MWReplyProxyFwd.h>
+#include <unity/scopes/ChildScope.h>
 #include <unity/scopes/Result.h>
 #include <unity/scopes/Variant.h>
 
@@ -66,6 +67,9 @@ public:
                                      ActionMetadata const& hints,
                                      MWReplyProxy const& reply,
                                      InvokeInfo const& info) = 0;
+
+    virtual ChildScopeList child_scopes_ordered() const = 0;
+    virtual bool set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered) = 0;
 
     virtual bool debug_mode() const = 0;
 };