+++ include/unity/scopes/internal/zmq_middleware/ZmqScope.h
@@ -62,6 +62,9 @@ public:
                                    VariantMap const& hints,
                                    MWReplyProxy const& reply) override;
 
+    virtual ChildScopeList child_scopes_ordered() override;
+    virtual bool set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered) override;
+
     virtual bool debug_mode() override;
 
 private: