+++ src/scopes/internal/ScopeImpl.cpp
@@ -283,6 +283,16 @@ QueryCtrlProxy ScopeImpl::preview(Result const& result,
     return ctrl;
 }
 
+ChildScopeList ScopeImpl::child_scopes_ordered()
+{
+    return fwd()->child_scopes_ordered();
+}
+
+bool ScopeImpl::set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered)
+{
+    return fwd()->set_child_scopes_ordered(child_scopes_ordered);
+}
+
 ScopeProxy ScopeImpl::create(MWScopeProxy const& mw_proxy, std::string const& scope_id)
 {
     return make_shared<ScopeImpl>(mw_proxy, scope_id);