+++ test/gtest/scopes/internal/zmq_middleware/ZmqMiddleware/ZmqMiddleware_test.cpp
@@ -453,6 +453,16 @@ public:
         return nullptr;
     }
 
+    virtual ChildScopeList child_scopes_ordered() const override
+    {
+        return ChildScopeList();
+    }
+
+    virtual bool set_child_scopes_ordered(ChildScopeList const&) override
+    {
+        return false;
+    }
+
     virtual bool debug_mode() const override
     {
         return false;