+++ RELEASE_NOTES.md
@@ -1,6 +1,13 @@
 Release notes
 =============
 
+Changes in version 0.6.12
+=========================
+  - Introduced child_scopes() methods for aggregators to return their
+    list of child scopes at runtime.
+  - Added missing virtual destructor to AbstractScopeBase. (LP: #1360266)
+  - Removed deprecated Runtime::run_scope() method.
+
 Changes in version 0.6.11
 =========================
   - The JSON for a CategoryRenderer now supports a "fallback" field in the